requirements.txt
accelerate>=1.1.1
aiofiles>=23.2.1
aiohappyeyeballs>=2.4.3
aiohttp>=3.11.7
aiosignal>=1.3.1
annotated-types>=0.7.0
anyascii>=0.3.2
anyio>=4.6.2.post1
attrs>=24.2.0
backoff>=2.2.1
bespokelabs-curator>=0.1.24
boto3>=1.34.162
botocore>=1.34.162
cachetools>=5.5.2
certifi>=2024.8.30
cffi>=1.17.1
charset-normalizer>=3.4.0
click>=8.1.7
cloudpickle>=3.1.0
compressed-tensors>=0.8.0
contourpy>=1.3.0
contractions>=0.1.73
cycler>=0.12.1
Cython>=3.0.11
datasets>=3.1.0
dill>=0.3.8
diskcache>=5.6.3
distro>=1.9.0
docstring_parser>=0.16
dyNET38>=2.2
einops>=0.8.0
eval_type_backport>=0.2.2
fastapi>=0.115.5
fastchat>=0.1.0
ffmpy>=0.4.0
filelock>=3.16.1
fonttools>=4.55.0
frozenlist>=1.5.0
fschat>=0.2.36
fsspec>=2024.9.0
gguf>=0.10.0
google-api-core>=2.24.2
google-auth>=2.40.1
google-cloud-aiplatform>=1.71.1
google-cloud-bigquery>=3.31.0
google-cloud-core>=2.4.3
google-cloud-resource-manager>=1.14.2
google-cloud-storage>=2.19.0
google-crc32c>=1.7.1
google-resumable-media>=2.7.2
googleapis-common-protos>=1.70.0
gradio_client>=1.3.0
grpc-google-iam-v1>=0.14.2
grpcio>=1.71.0
grpcio-status>=1.71.0
h11>=0.14.0
httpcore>=1.0.7
httptools>=0.6.4
httpx>=0.28.1
huggingface-hub>=0.26.2
idna>=3.10
instructor>=1.8.1
interegular>=0.3.3
Janome>=0.5.0
Jinja2>=3.1.4
jiter>=0.7.1
jmespath>=1.0.1
joblib>=1.4.2
jsonschema>=4.23.0
jsonschema-specifications>=2024.10.1
kiwisolver>=1.4.7
konoha>=5.5.6
kytea>=0.1.9
lark>=1.2.2
latex2mathml>=3.77.0
litellm>=1.61.3
llvmlite>=0.43.0
lm-format-enforcer>=0.10.9
markdown-it-py>=3.0.0
markdown2>=2.5.1
MarkupSafe>=2.1.5
matplotlib>=3.9.2
mdurl>=0.1.2
mistral_common>=1.5.1
mistralai>=1.7.0
monotonic>=1.6
mpmath>=1.3.0
msgpack>=1.1.0
msgspec>=0.18.6
multidict>=6.1.0
multiprocess>=0.70.16
nagisa>=0.2.11
natto-py>=1.0.1
nest-asyncio>=1.6.0
networkx>=3.2.1
nh3>=0.2.18
numba>=0.60.0
numpy>=1.26.4
openai>=1.78.0
opencv-python-headless>=4.10.0.84
orjson>=3.10.12
outlines>=0.0.46
pandas>=2.2.2
partial-json-parser>=0.2.1.1.post4
peft>=0.13.2
pillow>=10.4.0
posthog>=3.25.0
prometheus-fastapi-instrumentator>=7.0.0
prometheus_client>=0.21.0
prompt_toolkit>=3.0.51
propcache>=0.2.0
proto-plus>=1.26.1
psutil>=7.0.0
py-cpuinfo>=9.0.0
pyahocorasick>=2.1.0
pyairports>=2.1.1
pyarrow>=18.1.0
pyasn1>=0.6.1
pyasn1_modules>=0.4.2
pycountry>=24.6.1
pycparser>=2.22
pydantic>=2.11.4
pydantic_core>=2.33.2
pydub>=0.25.1
Pygments>=2.19.1
pyparsing>=3.2.0
python-dateutil>=2.9.0.post0
python-dotenv>=1.0.1
python-multipart>=0.0.12
pytz>=2024.2
PyYAML>=6.0.2
pyzmq>=26.4.0
ray>=2.39.0
referencing>=0.35.1
regex>=2024.11.6
requests>=2.32.3
rich>=13.9.4
rpds-py>=0.21.0
rsa>=4.9.1
ruff>=0.8.0
s3transfer>=0.10.4
safehttpx>=0.1.1
safetensors>=0.4.5
scikit-learn>=1.3.2
scipy>=1.13.1
semantic-version>=2.10.0
sentencepiece>=0.2.0
shapely>=2.1.0
shellingham>=1.5.4
shortuuid>=1.0.13
six>=1.17.0
sniffio>=1.3.1
starlette>=0.41.3
SudachiDict-core>=20230927
SudachiPy>=0.6.10
svgwrite>=1.4.3
sympy>=1.13.1
textsearch>=0.0.24
threadpoolctl>=3.5.0
tiktoken>=0.7.0
tokenizers>=0.20.3
tomlkit>=0.12.0
torch>=2.5.1
torchvision>=0.20.1
tqdm>=4.67.1
transformers>=4.46.3
triton>=3.1.0
typer>=0.13.1
typing-inspection>=0.4.0
typing_extensions>=4.13.2
tzdata>=2024.2
urllib3>=2.4.0
uvicorn>=0.25.0
uvloop>=0.21.0
vertexai>=1.71.1
vllm>=0.7.3
watchfiles>=1.0.0
wavedrom>=2.0.3.post3
wcwidth>=0.2.13
websockets>=12.0
xformers>=0.0.28.post3
xxhash>=3.5.0
yarl>=1.18.0
zipp>=3.21.0
langchain_anthropic
evaluate
bert_score
